Use SFINAE instead of macros for 'long' hack (#12605)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/12605

Some compilers define 'long' as a separate type from 'int32_t' or 'int64_t', some don't.
Before, we had a cmake check setting a macro and depending on the macro, we either defined a separate type id for long or didn't.
Then, we removed the cmake check and used compiler detection macros directly. This is, however, error prone.

This new approach uses SFINAE to register a type id for 'long' only if it's a separate type.

Reviewed By: Prowindy

Differential Revision: D10359443

fbshipit-source-id: aa371cbb43658c8cd3664ba3d9b0dedbaa225c1d