Remove net_simple_async (#13320)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/13320

simple_async has been deprecated via the network override rule for a while,
and we should be able to safely remove it.

This also clears up 2 tech debts:
(1) in rnn executor, rely on the executor override to get the right net.
(2) clearly mark checkExecutorOverride as a potential change to net_type by making it c++ style guide compliant.

Reviewed By: dzhulgakov

Differential Revision: D12840709

fbshipit-source-id: 667702045fa024f5bdc87a9c28ea1786c78432b3