Add c10::Stream, make at::cuda::CUDAStream use it. (#13133)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/13133

c10::Stream is a device agnostic object which represents a stream
on some device (defined as c10::Device).  The primary benefit of
introducing this object is that we can easily refer to it from code
in the non-CUDA library (since it doesn't actually refer to any
CUDA specific bits.)

Streams are identified by an ID into an appropriate pool.  There's
some work to translate to and from pointers to the pool; see inline
comments.

Reviewed By: gchanan

Differential Revision: D10855883

fbshipit-source-id: cc447f11a528432e41c2edc789f40e7a6f17bdd3