Do not fill in new data in every iteration if the input data only has one entry (#13495)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/13495

If the user has one data file to put in, the data is filled up in every iteration,
which actually flushes the caches. The retrieved latency is larger than the latency
when the caches are warm. Instead of doing that, we should only rely on wipe_cache
variable to wipe the caches.

The change is to skip filling the data if the input only has one size and it is
not the first iteration

Reviewed By: hl475

Differential Revision: D12897946

fbshipit-source-id: ee54ed09b8ec85fcefe930858420b90d494ad972